Expand the registrant form by default if there is only
The registrant form can and should be expanded by default if there is only one person type bundle. It saves one click (UX) on the button in this case.
If that should not happen by default in the view of the maintainer there should be at least an option to do this, because it's helpful in many cases where a quick simple registration form is required.